Update README.md
    Analyze sales data from more than 16,500 games.

Sales data from more than 16,500 games is interesting dataset to analyze the game's industry business side. As a person, who likes to play games it will be double interesting for me to analyze sales videogames data as well as this analysis could give ideas for investors to understand which games, genres, publishers can generate higher Global sales. As well as to understand the correlation between European and Japanese and North American markets.

The dataset used is called Video Game Sales and is available for free on Kaggle. Kaggle is a webservice that supplies a broad variation of datasets for free. The Formula 1 dataset consists of 13 different csv files. There is a separate file for drivers, constructors, lap times, pit stops, ... Some files can be linked together using id's.

This analysis is part of the course Data Analysis with Python: Zero to Pandas. This course starts with entry level Python programming and is thus available for everyone. The lessons are clearly explained and there is a wide community of people to help answer questions. If you are looking to get into Python I can highly recommend any of the courses listed here. Great thanks to Jovian for the (free) high quality lessons.